requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
pandas
tensorboard
fairscale
sentencepiece
fire
packaging
transformers
pyyaml
pathlib
Ninja